Update components with more reusable props
